Final smart health/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
albumentations>=1.1.0
matplotlib>=3.4.0
tensorboard>=2.7.0
scikit-image>=0.18.0
scipy>=1.7.0
tqdm>=4.62.0
PyYAML>=5.4.0
seaborn>=0.11.0
pandas>=1.3.0 